Move Flow types around, fix spelling error
